Strings are anagrams or not
Two strings A and B are called anagrams if they contain all the same characters in the same frequencies. Given two strings, find whether both are anagram of each other or not.
Complete the function isAnagram() that will take two strings as parameters and return 1 if they are anagram and return 0 otherwise.