Reverse the words of string
A string is composing some words. These words are separated by some delimiter e.g. a space.
Given a string (with space as a delimiter between words), reverse the words in the string individually, not the whole string.
Complete the function revWordsString() which takes the string as parameter and reverse the individual words of string and print them.